+++ src/main/java/blackjack/domain/card/Deck.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain.card;
 
 import java.util.ArrayList;
 import java.util.Arrays;
