+++ src/main/java/blackjack/domain/card/Rank.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain.card;
 
 public enum Rank {
     ONE(1), TWO(2), THREE(3), FOUR(4), FIVE(5),
@@ -9,5 +9,9 @@ public enum Rank {
 
     Rank(int value) {
         this.value = value;
+    }
+
+    public int getValue() {
+        return value;
     }
 }
