+++ src/main/java/blackjack/domain/card/Suit.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.domain.card;
 
 public enum Suit {
     SPACE, DIAMOND, HEART, CLOVER;
